intel_extreme: fix overflow in SKL PLL calculation
* RGB 4K@60Hz clock is 594MHz, that multiplied by 5 overflows int value.

Reviewed-on: https://review.haiku-os.org/c/haiku/+/4738
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ src/add-ons/accelerants/intel_extreme/pll.cpp
@@ -991,7 +991,7 @@ skl_ddi_calculate_wrpll(int clock /* in Hz */,
 			int ref_clock,
 			struct skl_wrpll_params *wrpll_params)
 {
-	uint64 afe_clock = clock * 5; /* AFE Clock is 5x Pixel clock */
+	uint64 afe_clock = (uint64) clock * 5; /* AFE Clock is 5x Pixel clock */
 	uint64 dco_central_freq[3] = { 8400000000ULL,
 				    9000000000ULL,
 				    9600000000ULL };
